+++ docs/develop/change-log.md
@@ -1,10 +1,12 @@
 # Change Log
+
 ### Version 2.4.0
 
-- improve: extract client and server common entity 
+- improve: extract client and server common entity
 - conf: **break change** remove hocon config
 - ci: add js client build ci, improve release cd
 - improve: clean deps, update claps
+- feat: add openTelemetry trace
 
 ### Version 2.3.0